move auto shard dataset checkpoint tests to kernel tests
